chore!: migrate to @grpc/grpc-js
This migrates the `grpc` package to the pure javascript `grpc-js`
implementation. `grpc-js` appears to be the more actively maintained
package and the preferred implementation by Google projects, and it
reduces the dependency on and overhead from compilers like
`node-pre-gyp`.

This also updates `grpc-tools` and tweaks the way grpc code is generated
from `.proto` files to be compatible with `grpc-js`.

Lastly this removes the grpc web proxy from xud. It did not appear to be
compatible with `grpc-js` and it is currently not used for any purpose
including testing or development. As the old approach relied on a custom
fork of an old, unmaintained package to create the proxy, should the need
arise for a web proxy in the future we can implement a more robust and
updated solution.

BREAKING CHANGE: Removes grpc web proxy from xud.